Remove http slack adaptor
